requirements.txt
# Core dependencies - install numpy first to avoid build issues
numpy>=1.24.0,<2.0.0

# LangChain and related
langchain>=0.1.0,<0.3.0
langchain-community>=0.0.13,<0.3.0
langchain-openai>=0.0.2,<0.2.0
langchain-core>=0.1.10,<0.3.0

# Vector store and embeddings
chromadb>=0.4.22,<0.6.0
sentence-transformers>=2.2.2,<3.0.0

# OpenAI
openai>=1.7.2,<2.0.0

# Web framework
flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0
